fix:move table component to shortlet index file
